Tentative de correction de tests
