+++ .docker/nginx/conf/franceconnect.gouv.test.conf
@@ -1,0 +1,40 @@
+server {
+    listen 443 ssl;
+    listen [::]:443 ssl;
+    http2  on;
+
+# SSL
+    ssl_certificate /etc/nginx/certs/franceconnect.gouv.test.pem;
+    ssl_certificate_key /etc/nginx/certs/franceconnect.gouv.test-key.pem;
+    ssl_trusted_certificate /etc/nginx/certs/rootCA.pem;
+
+    server_name franceconnect.gouv.test;
+
+    location / {
+        # Allow performing host detection on request
+        resolver 127.0.0.11 valid=1s;
+        set $upstream franceconnect:9997;
+
+            proxy_http_version  1.1;
+            proxy_cache_bypass  $http_upgrade;
+            proxy_set_header Host              $host;
+            proxy_set_header X-Forwarded-For   $proxy_add_x_forwarded_for;
+            proxy_set_header X-Real-IP         $remote_addr;
+            proxy_set_header X-Forwarded-Host  $host;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_set_header X-Forwarded-Port  $server_port;
+
+        proxy_pass http://$upstream;
+    }
+}
+
+server {
+    listen 80;
+    listen [::]:80;
+
+    server_name franceconnect.gouv.test;
+
+    location / {
+        return 301 https://franceconnect.gouv.test/$request_uri;
+    }
+}
