+++ src/Security/Oidc/OidcClient.php
@@ -116,8 +116,12 @@ class OidcClient
         );
     }
 
-    public function buildLogoutUrl(Request $request, string $idToken): string
+    public function buildLogoutUrl(Request $request, string $idToken): ?string
     {
+        if (!isset($this->configuration['end_session_endpoint'])) {
+            return null;
+        }
+
         $state = Uuid::uuid4()->toString();
         $redirectUri = $this->urlGenerator->generate($this->logoutRoute, referenceType: UrlGeneratorInterface::ABSOLUTE_URL);
